Fix build

+++ src/battle_anim_utility_funcs.c
@@ -768,7 +768,7 @@ void sub_81177E4(u8 taskId)
 }
 
 
-void sub_8117854(u8 taskId, int unused, u16 arg2, u8 battler1, u8 arg4, u8 arg5, u8 arg6, u8 arg7, const u8 *gfx, const u8 *tilemap, const u16 *palette)
+void sub_8117854(u8 taskId, int unused, u16 arg2, u8 battler1, u8 arg4, u8 arg5, u8 arg6, u8 arg7, const u32 *gfx, const u32 *tilemap, const u32 *palette)
 {
     u16 species;
     u8 spriteId, spriteId2;
